Disable Mpris
I think it (well, the way I used it) was the cause of some crashes I
was experiencing; after disabling, I've never experienced them again.

It also didn't work for mopidy, which I think works with a different
DBus session.

It might be worth some time to investigate this further, because I
really like the idea.